Conflicts:
	file/tour.exerd

+++ src/main/java/com/yi/tourland/controller/mng/ManagerController.java
@@ -715,16 +715,18 @@ public class ManagerController {
 		return "manager/product/productModify";
 	}
 	@RequestMapping(value = "productModify", method = RequestMethod.POST)
-	public String productModifyPost(int[] airNo, int[] hotelNo, int[] tourNo, int[] rentcarNo, ProductVO vo, MultipartFile file, SearchCriteria cri,Model model) throws Exception {
+	public String productModifyPost(int[] airNo, int[] hotelNo, int[] tourNo, int[] rentcarNo, ProductVO vo, MultipartFile file, SearchCriteria cri,String originalPic,Model model) throws Exception {
 		int pno = vo.getPno();
 		ProductVO delProd = new ProductVO(pno);
 		productService.deleteProduct(delProd);
-		if (vo.getPic() != null) {
+		System.out.println(file.getSize());
+		System.out.println(vo.getPic());
+		if (file.getSize()>0 && vo.getPic()!=null) {
 			File pBig = new File(uploadPathProduct + vo.getPic());
-			pBig.delete();
+			if(pBig.exists()) pBig.delete();
 			String updatePath = vo.getPic().substring(0, 12) + "s_" + vo.getPic().substring(12);
 			File pSmall = new File(uploadPathProduct + updatePath);
-			pSmall.delete();
+			if(pSmall.exists()) pSmall.delete();
 		}
 		List<AirplaneVO> air = new ArrayList<>();
 		List<HotelVO> hotel = new ArrayList<>();
@@ -750,12 +752,17 @@ public class ManagerController {
 		vo.setHotel(hotel);
 		vo.setTour(tour);
 		vo.setRentcar(rentcar);
-		File filePath = new File("uploadPathProduct");
-		if(!filePath.exists()) filePath.mkdir();
-		String savedName = UploadFileUtils.uploadFile(uploadPathProduct, file.getOriginalFilename().replaceAll(" ", "_"),
-				file.getBytes());
-		String bigSizePic = savedName.substring(0, 12) + savedName.substring(14);
-		vo.setPic(bigSizePic.replaceAll(" ", "_"));
+		if(file.getSize()>0 && vo.getPic()==null) {
+			File filePath = new File("uploadPathProduct");
+			if(!filePath.exists()) filePath.mkdir();
+			String savedName = UploadFileUtils.uploadFile(uploadPathProduct, file.getOriginalFilename().replaceAll(" ", "_"),
+					file.getBytes());
+			String bigSizePic = savedName.substring(0, 12) + savedName.substring(14);
+			vo.setPic(bigSizePic.replaceAll(" ", "_"));
+		}
+		else {
+			vo.setPic(originalPic);
+		}
 		productService.insertProduct(vo);
 		return "redirect:productDetail?no="+vo.getPno()+"&page="+cri.getPage()+"&searchType="+cri.getSearchType()+"&keyword="+cri.getKeyword();
 	}
@@ -910,6 +917,7 @@ public class ManagerController {
 			PageMaker pageMaker = new PageMaker();
 			pageMaker.setCri(cri);
 			pageMaker.setTotalCount(rentcarService.totalSearchCountRentcar(cri));
+		
 			map.put("list", rentList);
 			map.put("pageMaker", pageMaker);
 			entity = new ResponseEntity<Map<String,Object>>(map, HttpStatus.OK);
@@ -1006,7 +1014,12 @@ public class ManagerController {
 		pageMaker.setTotalCount(rentcarService.totalSearchCountRentcar(cri));
 		
 	//	System.out.println(rentcarService.totalSearchCountRentcar(cri));
-
+		System.out.println("리스트 사이즈"+rentcarList.size());
+		for(RentcarVO rentcar: rentcarList) {
+			  System.out.println("렌트"+rentcar.getRentddate());
+			  System.out.println("리턴"+rentcar.getReturndate());
+			  
+		}
 		model.addAttribute("cri", cri);
 		model.addAttribute("list", rentcarList);
 		model.addAttribute("pageMaker", pageMaker);
@@ -1105,7 +1118,6 @@ public class ManagerController {
 	 
 	 @RequestMapping(value = "eventRegister", method = RequestMethod.POST)
 		public String eventRegisterPost(EventVO vo, MultipartFile eventPic, Model model) throws Exception {
-
 			String savedName = UploadFileUtils.uploadFile(uploadPathEvent, eventPic.getOriginalFilename().replaceAll(" ", "_"),
 					eventPic.getBytes());
 			String bigSizePic = savedName.substring(0, 12) + savedName.substring(14);
@@ -1126,7 +1138,7 @@ public class ManagerController {
 		
 		@RequestMapping(value = "eventUpdate", method = RequestMethod.POST)
 		public String eventUpdate(EventVO vo, MultipartFile eventPic, Model model) throws Exception {
-		
+		     System.out.println(vo);
 			if (eventPic.getBytes().length != 0) { // 새로 첨부한 파일이 있다면
 				// 원래 vo가 가진 pic의 네임으로 폴더에 저장된 사진들 지우기
